requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cloudinary==1.33.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==4.1.0
django-environ==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
newsdataapi==0.1.6
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.6.3
tzlocal==5.0.1
urllib3==1.26.16
